Continuing the integer datatype conversion.
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@249 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/d_auto_filter.c
@@ -421,7 +421,7 @@ mxd_auto_filter_change_control( MX_AUTOSCALE *autoscale )
 	static const char fname[] = "mxd_auto_filter_change_control()";
 
 	MX_AUTO_FILTER *auto_filter;
-	unsigned long old_filter_setting, new_filter_setting;
+	uint32_t old_filter_setting, new_filter_setting;
 	mx_status_type mx_status;
 
 	mx_status = mxd_auto_filter_get_pointers( autoscale,
@@ -474,7 +474,8 @@ mxd_auto_filter_change_control( MX_AUTOSCALE *autoscale )
 
 	MX_DEBUG( 2,("%s: Changing filter '%s' from %#lx to %#lx.",
 		fname, autoscale->control_record->name,
-		old_filter_setting, new_filter_setting));
+		(unsigned long) old_filter_setting,
+		(unsigned long) new_filter_setting));
 
 	mx_status = mx_digital_output_write( autoscale->control_record,
 						new_filter_setting );
@@ -502,7 +503,7 @@ mxd_auto_filter_set_offset_index( MX_AUTOSCALE *autoscale )
 {
 	static const char fname[] = "mxd_auto_filter_set_offset_index()";
 
-	unsigned long saved_index;
+	uint32_t saved_index;
 
 	/* The offset index for the autoscale filter driver is only
 	 * allowed to be zero.
@@ -517,7 +518,7 @@ mxd_auto_filter_set_offset_index( MX_AUTOSCALE *autoscale )
 		return mx_error( MXE_ILLEGAL_ARGUMENT, fname,
 		"The requested monitor offset index of %lu is not allowed.  "
 		"Zero is the only allowed value.",
-			saved_index );
+			(unsigned long) saved_index );
 	}
 
 	return MX_SUCCESSFUL_RESULT;
